+++ opentelemetry-sdk/src/opentelemetry/sdk/trace/__init__.py
@@ -424,6 +424,12 @@ class Span(trace_api.Span):
             # Freeze mutable sequences defensively
             if isinstance(value, MutableSequence):
                 value = tuple(value)
+            if isinstance(value, bytes):
+                try:
+                    value = value.decode()
+                except ValueError:
+                    logger.warning("Byte attribute could not be decoded.")
+                    return
             with self._lock:
                 self.attributes[key] = value